+++ packages/mcp-provider-dx-core/CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### Bug Fixes
 
-* improve sf-get-username instructions W-19472383 ([#181](https://github.com/salesforcecli/mcp/issues/181)) ([ce845f8](https://github.com/salesforcecli/mcp/commit/ce845f8a956dece92cf2d67e29cc868c8ac69b92))
+* improve get_username instructions W-19472383 ([#181](https://github.com/salesforcecli/mcp/issues/181)) ([ce845f8](https://github.com/salesforcecli/mcp/commit/ce845f8a956dece92cf2d67e29cc868c8ac69b92))